requirements.txt
streamlit
kafka-python
feedparser
transformers
torch
requests
